minor shader enhancements

+++ Intern/RayCore/src/Constants.h
@@ -49,13 +49,13 @@ const double W_UNINIT = 4;
 // -----------------------------------------
 
 // The meaning of surfaceParams[3][0] (i.e. the type of an element):
-const double TY_BASIC_MIRROR = 0;  // plane + sphere mirror
-const double TY_PLANE_GRATING = 1;
-const double TY_GENERAL_GRATING = 2;
-const double TY_SLIT = 3;
-const double TY_RZP = 4;
-const double TY_IMAGE_PLANE = 5;
-const double TY_TOROID_MIRROR = 6;
-const double TY_ELLIPSOID_MIRROR = 7;
+const int TY_BASIC_MIRROR = 0;  // plane + sphere mirror
+const int TY_PLANE_GRATING = 1;
+const int TY_GENERAL_GRATING = 2;
+const int TY_SLIT = 3;
+const int TY_RZP = 4;
+const int TY_IMAGE_PLANE = 5;
+const int TY_TOROID_MIRROR = 6;
+const int TY_ELLIPSOID_MIRROR = 7;
 
 #endif
